fix: convert retry-builds into cloud run job
